requirements.txt
google-generativeai>=0.8.3
pandas
playwright
streamlit
plotly